# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Microservice to add reviews to articles",
        "title": "Reviews Microservice",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/articles/{id}/rating": {
            "get": {
                "description": "Get the rating of an article",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reviews"
                ],
                "summary": "Rating for an article",
                "parameters": [
                    {
                        "type": "string",
                        "description": "article id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/vo.ReviewsRating"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrCustom"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrCustom"
                        }
                    }
                }
            }
        },
        "/articles/{id}/reviews": {
            "get": {
                "description": "Get the reviews of an article",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reviews"
                ],
                "summary": "List of reviews for an article",
                "parameters": [
                    {
                        "type": "string",
                        "description": "article id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/review.Review"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrCustom"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrCustom"
                        }
                    }
                }
            }
        },
        "/reviews": {
            "post": {
                "description": "Add a review for an article",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reviews"
                ],
                "summary": "Add a review",
                "parameters": [
                    {
                        "description": "Add review",
                        "name": "review",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/vo.CreateReview"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/vo.CreateReview"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrCustom"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrCustom"
                        }
                    }
                }
            }
        },
        "/reviews/{id}": {
            "delete": {
                "description": "Delete a review by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reviews"
                ],
                "summary": "Delete a review",
                "parameters": [
                    {
                        "type": "string",
                        "description": "review id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrCustom"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrCustom"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "errors.ErrCustom": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "review.Review": {
            "type": "object",
            "properties": {
                "article_id": {
                    "type": "string"
                },
                "comment": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "score": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "vo.CreateReview": {
            "type": "object",
            "properties": {
                "article_id": {
                    "type": "string"
                },
                "comment": {
                    "type": "string"
                },
                "score": {
                    "type": "integer"
                }
            }
        },
        "vo.ReviewsRating": {
            "type": "object",
            "properties": {
                "article_id": {
                    "type": "string"
                },
                "rating": {
                    "type": "number"
                }
            }
        }
    }
}